requirements.txt
websockets==12.0
httpx==0.27.0
openai==1.51.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pyyaml==6.0.1
loguru==0.7.2
python-dotenv==1.0.1
uvloop==0.19.0
